Move Dfs and so on into their own file under visit
